RoleRepository and UserRepository created

+++ src/main/java/com/arab/assessment/entities/Role.java
@@ -13,7 +13,7 @@ public class Role {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	private Integer id;
 	
 	@Column(nullable = false, length = 20)
 	private String name;
@@ -24,20 +24,20 @@ public class Role {
 		this.name = name;
 	}
 	
-	public Role(Long id) {
+	public Role(Integer id) {
 		this.id = id;
 	}
 	
-	public Role(Long id, String name) {
+	public Role(Integer id, String name) {
 		this.id = id;
 		this.name = name;
 	}
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
